+++ examples/demo/config/dev.exs
@@ -20,7 +20,7 @@ config :demo, DemoWeb.Endpoint,
   check_origin: false,
   code_reloader: true,
   debug_errors: true,
-  secret_key_base: "vZGdHWjCQXGUd72lGxw6OGN9zuL1Y3r46My0HdqNvRacEB2ItBFJQ0bN/9nh/z+q",
+  secret_key_base: "TxTvvr819X//+gdsL7KlGHrQKqLjTUduVWTFQFlJD+0/LJuoEOjTOWWJH+5iKKmh",
   watchers: [
     esbuild: {Esbuild, :install_and_run, [:demo, ~w(--sourcemap=inline --watch)]},
     tailwind: {Tailwind, :install_and_run, [:demo, ~w(--watch)]}